pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["play_outside"]

[project]
name = "play-outside"
dynamic = ["version"]
description = 'Use Open Weather Api to determine if Kids can play outside.'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [{ name = "Waylon S. Walker", email = "waylon@waylonwalker.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    'httpx',
    'fastapi',
    'uvicorn[standard]',
    'typer',
    'python-dotenv',
    'rich',
    'pydantic-settings',
    'jinja2',
    'arel',
]


[project.urls]
Documentation = "https://github.com/unknown/play-outside#readme"
Issues = "https://github.com/unknown/play-outside/issues"
Source = "https://github.com/unknown/play-outside"

[project.scripts]
play-outside = "play_outside.cli.api:api_app"

[tool.hatch.version]
path = "play_outside/__about__.py"

[tool.hatch.envs.default]
dependencies = ["coverage[toml]>=6.5", "pytest"]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = ["- coverage combine", "coverage report"]
cov = ["test-cov", "cov-report"]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:play_outside tests}"

[tool.coverage.run]
source_pkgs = ["play_outside", "tests"]
branch = true
parallel = true
omit = ["play_outside/__about__.py"]

[tool.coverage.paths]
play_outside = ["play_outside", "*/play-outside/play_outside"]
tests = ["tests", "*/play-outside/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
